+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Next version
 + Add scss
++ Save expire token date
 
 ## 0.11.0
 + Move router to separate component